Feat: "DashBoard에 사용되는 Repository에 MongoDB Aggregation 적용."

+++ src/main/java/com/jungwoo/apiserver/dto/mongo/detect/DashboardDto.java
@@ -14,9 +14,9 @@ import java.util.List;
 @Builder
 public class DashboardDto {
 
-  private Long keywordCount;
+  private Integer keywordCount;
 
-  private Long memberCount;
+  private Integer memberCount;
 
   private Double averageAccuracy;
